requirements.txt
asgiref==3.2.3
autopep8==1.5
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Django==3.0.5
django-mongodb-engine==0.6.0
djangorestframework==3.11.0
djangotoolbox==1.8.0
idna==2.9
numpy==1.18.2
pycodestyle==2.5.0
pymongo==3.10.1
PyQRCode==1.2.1
pytz==2019.3
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.0
urllib3==1.25.8
virtualenv==16.7.9
